Create contacts, calendar, files api functions
I implemented the contacts, calendar and file google apis.  I am in the process
of outputing the contacts of a user properly.  For the calendars api, I am able
to get the json for the calendar events of a user's primary calendar.  I will
now parse through the events in the calendar and find the events that are
relevant for the current week.