Complete week events extractor
I was able to extract the events for the current week from a user's calendar in
the getCurrentWeekCalendarEvents method.  It uses the json simple and calendar
libraries to search through the json object returned from the calendar api and
extract the events of the calendar which fall between Monday and Sunday of this
week.